pyproject.toml
[project]
name = "woeplanet-spelunker"
version = "2.0.0"
description = "The WOEplanet Spelunker"
authors = [{name="Gary Gale", email="gary@vicchi.org"}]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "aiosqlite>=0.21.0",
    "aiosqlitepool>=1.0.0",
    "diskcache>=5.6.3",
    "emoji>=2.15.0",
    "inflect>=7.5.0",
    "jinja2>=3.1.6",
    "pycountry>=24.6.1",
    "pydantic>=2.12.4",
    "pydantic-settings>=2.12.0",
    "python-dotenv>=1.2.1",
    "pyuca>=1.2",
    "pyyaml>=6.0.3",
    "starlette>=0.50.0",
    "starlette-async-jinja>=1.13.3",
    "uvicorn>=0.38.0",
]

[project.scripts]
server = "woeplanet.spelunker.server:main"

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "mypy>=1.18.2",
    "parametrize-from-file>=0.20.0",
    "pyproject-parser[cli]>=0.13.0",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.3.0",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.5",
    "types-pyyaml>=6.0.12.20250915",
    "yamllint>=1.37.1",
]

[build-system]
requires = ["uv_build>=0.9.9,<0.10.0"]
build-backend = "uv_build"

[tool.mypy]
mypy_path = "src"
explicit_package_bases = true
plugins = ['pydantic.mypy']

[[tool.mypy.overrides]]
module="aiosqlitepool"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module="parametrize_from_file"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module="pyuca"
ignore_missing_imports = true

[[tool.mypy.overrides]]
module="starlette_async_jinja"
ignore_missing_imports = true

[tool.ruff]
line-length = 120

[tool.ruff.format]
quote-style = "single"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.ruff.lint]
select = ["ALL"]
ignore=["COM812","D102","D104","D107","D200","D202","D203","D212","D400","D401","D415"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "single"

[tool.ruff.lint.per-file-ignores]
"tests/**/*.py" = ["S101"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
asyncio_default_fixture_loop_scope = "function"
addopts = "-p no:unraisableexception"
filterwarnings = [
    "ignore:unclosed database:ResourceWarning",
]

[tool.uv.build-backend]
module-name="woeplanet.spelunker"
